wip(test_signals): split long running signal tests
